--- a/Makefile
+++ b/Makefile
@@ -58,6 +58,7 @@ lint:
 functional-tests-exitfirst:
 ifeq ($(SLOW),)
 	echo "Running fast tests"
+	# `splits` and the `pytest-group` matrix on CI should match
 	pytest --exitfirst --splits 3 --group $(PYTEST_GROUP) --store-durations --durations-path=./.test_durations.$(PYTEST_GROUP) --deselect=tests/functional/targets/test_performance.py::test_api_requester_multiple_request_and_targets[50-50-600] tests -vvv --cucumberjson=test-report.json --durations=0 --html=test-report.html
 else
 	echo "Running slow tests"