--- a/tests/functional/scripts/run-ft-das.sh
+++ b/tests/functional/scripts/run-ft-das.sh
@@ -1,6 +1,9 @@
 #!/bin/bash -x
 
 
+PYTEST_GROUP=$2
+SLOW=$3
+
 # Base FT
 . "${UMBRELLA_PATH}/tests/functional/scripts/ft-base.sh"
 
@@ -87,4 +90,4 @@ if [[ ${UMBRELLA_PATH} != "." ]]; then
     cp metadata-update-payload.json ${UMBRELLA_PATH}/
 fi
 
-make -C ${UMBRELLA_PATH}/ functional-tests-exitfirst
+make -C ${UMBRELLA_PATH}/ functional-tests-exitfirst PYTEST_GROUP=${PYTEST_GROUP} SLOW=${SLOW}