### Description

Using `pytest-split` we can make the test duration of Functional tests that run on CI smaller, thus improving the development experience.

For Functional tests that we run from `repository-service-tuf-api`:
~6 min total duration for all Functional Tests to complete ([link](https://github.com/repository-service-tuf/repository-service-tuf-api/actions/runs/7602466421)) vs ~11-15 min (there's quite some deviation)

#### Improvements
There's still room for improvement by saving the test duration files, so that `pytest-split` can split the tests in better groups.
We can also flag the slow tests.

#### Testing it out & Performance
[Umbrella](https://github.com/repository-service-tuf/repository-service-tuf):
Sample run-times:
  - Before: [9m 51s](https://github.com/repository-service-tuf/repository-service-tuf/actions/runs/7159210504/usage)
  - After: [5m 15s](https://github.com/repository-service-tuf/repository-service-tuf/actions/runs/7603429978/usage)

Before merging we should already have prepared the CI for all the other services under the umbrella:
- [ ] API: https://github.com/repository-service-tuf/repository-service-tuf-api/pull/519 ([workflow run](https://github.com/repository-service-tuf/repository-service-tuf-api/actions/runs/7603779717))
Sample run-times:
  - Before: [10m 31s](https://github.com/repository-service-tuf/repository-service-tuf-api/actions/runs/7117579375/usage)
  - After: [5m 59s](https://github.com/repository-service-tuf/repository-service-tuf-api/actions/runs/7603779717/usage)
- [ ] Worker: https://github.com/repository-service-tuf/repository-service-tuf-worker/pull/440 - I don't have access to trigger and test it (see [here](https://github.com/repository-service-tuf/repository-service-tuf-worker/actions/workflows/review-approved.yml))
Sample run-times:
  - Before: [19m 46s](https://github.com/repository-service-tuf/repository-service-tuf-worker/actions/runs/7260720862/usage)
  - After: N/A
- [ ] CLI: https://github.com/repository-service-tuf/repository-service-tuf-cli/pull/492 ([workflow run](https://github.com/repository-service-tuf/repository-service-tuf-cli/actions/runs/7603487240))
Sample run-times:
  - Before: [9m 36s](https://github.com/repository-service-tuf/repository-service-tuf-cli/actions/runs/7263272330/usage)
  - After: [5m 4s](https://github.com/repository-service-tuf/repository-service-tuf-cli/actions/runs/7603487240/usage)

_Note for review: Check that in the end all tests are run after being split_

### Diagram

```mermaid
flowchart TB
    %% A sub-repo such as RSTUF-CLI
    subgraph "RSTUF sub-repo CI (e.g., CLI)"
    
    functional-tests

    %% A sub-repo such as RSTUF-CLI
    functional-tests-slow
    end

    %% Slow tests are ran separately
    functional-tests-slow -->|Run| slow-ci-job[Slow CI job]
    
    functional-tests -->|Checkout RSTUF umbrella| ci-job-group-one[CI job]
    functional-tests -->|Run|ci-job-group-two[CI job]
    functional-tests -->|Run| ci-job-group-three[CI job]

    ci-job-group-one -->|Run with latest RSTUF services| docker-services[Docker services]
    ci-job-group-one -->|SLOW: bool, GROUP: int - parameters passed| pytest[Run tests]

    rstuf-umbrella(Checked out RSTUF umbrella) -.->|Functional tests| pytest
    pytest -.-> docker-services
```