@KAUTH 2024-01-29
> I need introduction to these things and that's why I asked a lot of questions.

Sure, ask me anything that's not clear 👍