@ivanayov 2024-02-16
Bringing it here, as we discussed, so it's not lost in threads.

Is there a strong reason behind using an absolute order? 
Isn't it better to use `least_duration` instead, as it's more effective over time and would require less support (like changing order manually to improve performance)?