@KAUTH 2024-02-28
> Could you add some documentation about the tests and how we run them:
https://repository-service-tuf.readthedocs.io/en/stable/devel/development.html#functional-tests-using-behavior-development-driven-bdd

Added the relevant documentation

> Also, how to include a new slow tests and what is the best practises with pytest-split?
It will be good to mention that if we add a new slow test with new group or job, then we would need to change all YML files in all repositories.

I would include more details as a follow-up task after this is merged in all repos. It would be a good opportunity to explain better how our test suite is structured.