@KAUTH 2024-02-28
With the changes in the `Makefile` running locally might need to run `make` manually with different arguments in order to run all test sub-groups. We can improve this later on to be able to choose to run everything at once.